Move make_metadata into utils
