+++ xena_gdc_etl/main.py
@@ -1,6 +1,7 @@
 import argparse
 
-from .utils import equal_matrices
+from .utils import equal_matrices, metadata
+from .constants import valid_dtype
 
 
 def main():
@@ -12,6 +13,9 @@ def main():
     # handle check xena equality matrices
     if 'df1' in options and 'df2' in options:
         equal_matrices(options['df1'], options['df2'])
+    # handle make metadata
+    elif "matrix" in options and "datatype" in options:
+        metadata(options["matrix"], options["datatype"])
 
 
 def create_parser():
@@ -34,5 +38,18 @@ def create_parser():
     equality_parser.add_argument(
         "df2", type=str,
         help='Directory for the second matrix.'
+    )
+    make_metadata_parser = subparsers.add_parser(
+        "make-metadata",
+        help="Generating Xena metadata for a Xena matrix."
+    )
+    make_metadata_parser.add_argument(
+        '-m', '--matrix', type=str, required=True,
+        help='The path, including the filename, of the Xena matrix.'
+    )
+    make_metadata_parser.add_argument(
+        '-t', '--datatype', type=str, required=True,
+        help='One data type code indication the data type in matrices to be '
+        'merged. Supported data type codes include: {}'.format(str(valid_dtype))
     )
     return parser
